implements InventoryReport functionality

+++ src/main/java/com/moresby/ed/stockportfolio/inventory/Inventory.java
@@ -31,7 +31,7 @@ public class Inventory {
     private Integer amount;
 
     @Column(name = "cost")
-    private Double cost;
+    private Long cost;
 
     @ManyToOne
     @JoinColumn(
